backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
pandas
gunicorn
oracledb